refactor(search): downsize UnifiedSearchParams.top_k to u16 + timeout_ms to u32 (TD-SEARCH-2 S3)
Per-query thriftiness for the request struct (TD-SEARCH-2 S3 numeric slice):

- top_k: Option<usize> (16 B) -> Option<u16> (4 B). top_k is a small result
  count (realistic <= ~10^4, well under u16's 65_535 cap).
- timeout_ms: Option<u64> (16 B) -> Option<u32> (8 B). 4B ms (~49 days) is far
  beyond any query timeout.

UnifiedSearchParams is allocated/passed on every search, so this trims the
per-query struct. The small VALUE is preserved end-to-end; only the STORAGE
type shrinks.

Design: top_k is u16 in the request struct. Where it is read into engine
EXECUTION and used as a SIZE (BinaryHeap capacity, Vec::truncate, loop bound,
threshold), the local is usize and one `u16 as usize` cast sits at that
genuine size-boundary — usize is the correct Rust type for a size, and the cast
is transient stack (zero per-query cost). Set sites from usize sources cast
`as u16`. The ~100-file engine plumbing keeps top_k as usize (it is used as a
size there); u16 in the request struct is the per-query win.

core::search::SearchParams is a type alias for UnifiedSearchParams (one type).
Not on the OpenAPI spec (no ToSchema); serde ignores unknown fields, so the
wire is mixed-read-safe.

38 files: 2 field defs + Default in search_params.rs, the rest are boundary
casts at direct callers (set: `as u16`; size-read: `as usize`; timeout: `as u64`).

Verification (local, pre-CI):
- cargo check --all-targets (lib + tests + benches) OK
- cargo clippy --lib --bins (CI gate) OK
- cargo fmt --check OK
- cargo nextest -p proximadb-search-types: 40/40 pass (incl. serde round-trip)

Follow-up (separate): cascade u16 into other per-query structs (SearchContext,
SearchPlan, QueryContext) that still carry top_k: usize, to remove the residual
boundary casts.

+++ crates/foundation/proximadb-search-types/src/search_params.rs
@@ -431,8 +431,12 @@ pub struct UnifiedSearchParams {
     /// Single query vector (alternative to query_vectors for single queries)
     pub vector: Option<Vec<f32>>,
 
-    /// Number of results to return
-    pub top_k: Option<usize>,
+    /// Number of results to return. `u16` — top_k is a small result count
+    /// (realistic ≤ ~10⁴, well under u16's 65 535 cap), and this shrinks the
+    /// per-query field from `Option<usize>` (16 B) to `Option<u16>` (4 B).
+    /// Cast to `usize` only at genuine size-boundaries (Vec::truncate, heap
+    /// capacity, loop bounds); TD-SEARCH-2 S3.
+    pub top_k: Option<u16>,
 
     /// Distance metric to use for similarity calculation
     #[serde(skip)]
@@ -450,8 +454,10 @@ pub struct UnifiedSearchParams {
     /// Include expired vectors in results
     pub include_expired: Option<bool>,
 
-    /// Search timeout in milliseconds
-    pub timeout_ms: Option<u64>,
+    /// Search timeout in milliseconds. `u32` (not `u64`) — 4B ms ≈ 49 days is
+    /// far beyond any query timeout; halves the field (Option<u64> 16B →
+    /// Option<u32> 8B). Convert at use sites (`as u64`); TD-SEARCH-2 S3.
+    pub timeout_ms: Option<u32>,
 
     /// Enable two-stage search with quantization
     pub enable_two_stage: Option<bool>,
