+++ tests/sst_ivf2_compaction_route_proof_test.rs
@@ -152,7 +152,7 @@ async fn search_ids(engine: &SstEngine, coll: &Collection, query: Vec<f32>) -> V
     let ctx = StorageQueryContext {
         search_params: Arc::new(SearchParams {
             query_vectors: Some(vec![query]),
-            top_k: Some(TOP_K),
+            top_k: Some(TOP_K as u16),
             distance_metric: Some(DistanceMetric::Euclidean),
             block_prune: BlockPruneConfig {
                 radius_k: 0.0,
